pyproject.toml
[project]
name = "line-calendar-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "line-bot-sdk>=3.22,<4",
    "anthropic>=0.42,<1",
    "google-api-python-client>=2.160,<3",
    "google-auth>=2.37,<3",
    "google-auth-oauthlib>=1.2,<2",
    "google-cloud-firestore>=2.19,<3",
    "pydantic-settings>=2.7,<3",
    "cryptography>=44,<45",
    "httpx>=0.28,<1",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
]
